fix: refactor block config
