+++ frontend/appflowy_tauri/src/appflowy_app/stores/reducers/document/async-actions/blocks/index.ts
@@ -1,0 +1,1 @@
+export * from './text';
