+++ frontend/appflowy_tauri/src/appflowy_app/stores/reducers/document/async-actions/blocks/text/indent.ts
@@ -1,7 +1,7 @@
-import { BlockType, DocumentState } from '$app/interfaces/document';
+import { DocumentState } from '$app/interfaces/document';
 import { DocumentController } from '$app/stores/effects/document/document_controller';
 import { createAsyncThunk } from '@reduxjs/toolkit';
-import { allowedChildrenBlockTypes } from '$app/constants/document/config';
+import { blockConfig } from '$app/constants/document/config';
 
 export const indentNodeThunk = createAsyncThunk(
   'document/indentNode',
@@ -20,7 +20,8 @@ export const indentNodeThunk = createAsyncThunk(
     const newParentId = children[index - 1];
     const prevNode = state.nodes[newParentId];
     // check if prev node is allowed to have children
-    if (!allowedChildrenBlockTypes.includes(prevNode.type)) return;
+    const config = blockConfig[prevNode.type];
+    if (!config.canAddChild) return;
     // check if prev node has children and get last child for new prev node
     const prevNodeChildren = state.children[prevNode.children];
     const newPrevId = prevNodeChildren[prevNodeChildren.length - 1];
