+++ frontend/appflowy_tauri/src/appflowy_app/stores/reducers/document/async-actions/index.ts
@@ -1,7 +1,3 @@
-export * from './blocks/text/delete';
-export * from './blocks/text/indent';
-export * from './blocks/text/insert';
-export * from './blocks/text/backspace';
-export * from './blocks/text/outdent';
-export * from './blocks/text/split';
 export * from './cursor';
+export * from './blocks';
+export * from './turn_to';
