+++ frontend/appflowy_tauri/src/appflowy_app/stores/reducers/document/async-actions/turn_to.ts
@@ -2,9 +2,17 @@ import { createAsyncThunk } from '@reduxjs/toolkit';
 import { DocumentController } from '$app/stores/effects/document/document_controller';
 import { BlockData, BlockType, DocumentState, NestedBlock } from '$app/interfaces/document';
 import { setCursorBeforeThunk } from '$app_reducers/document/async-actions/cursor';
-import { allowedChildrenBlockTypes } from '$app/constants/document/config';
+import { blockConfig } from '$app/constants/document/config';
 import { newBlock } from '$app/utils/document/blocks/common';
 
+/**
+ * transform to block
+ * 1. insert block after current block
+ * 2. move all children
+ *    - if new block is not allowed to have children, move children to parent
+ *    - otherwise, move children to new block
+ * 3. delete current block
+ */
 export const turnToBlockThunk = createAsyncThunk(
   'document/turnToBlock',
   async (payload: { id: string; controller: DocumentController; type: BlockType; data: BlockData<any> }, thunkAPI) => {
@@ -18,19 +26,14 @@ export const turnToBlockThunk = createAsyncThunk(
     const parent = state.nodes[node.parent];
     const children = state.children[node.children].map((id) => state.nodes[id]);
 
-    /**
-     * transform to block
-     * 1. insert block after current block
-     * 2. move all children
-     * 3. delete current block
-     */
-
     const block = newBlock<any>(type, parent.id, data);
     // insert new block after current block
     const insertHeadingAction = controller.getInsertAction(block, node.id);
 
+    // check if prev node is allowed to have children
+    const config = blockConfig[block.type];
     // if new block is not allowed to have children, move children to parent
-    const newParent = allowedChildrenBlockTypes.includes(block.type) ? block : parent;
+    const newParent = config.canAddChild ? block : parent;
     // if move children to parent, set prev to current block, otherwise the prev is empty
     const newPrev = newParent.id === parent.id ? block.id : '';
     const moveChildrenActions = controller.getMoveChildrenAction(children, newParent.id, newPrev);
