--- frontend/appflowy_tauri/src/appflowy_app/utils/document/blocks/quote.ts
@@ -1,11 +1,0 @@
-import { Editor } from 'slate';
-import { getDeltaAfterSelection } from '$app/utils/document/blocks/common';
-
-export function getQuoteDataFromEditor(editor: Editor) {
-  const delta = getDeltaAfterSelection(editor);
-  if (!delta) return;
-  return {
-    delta,
-    size: 'default',
-  };
-}
